I warning di compilazione tornano a zero, e restano visibili
Il progetto compilava in silenzio per il motivo sbagliato: `-Xlint` non era
acceso, quindi javac non diceva niente e nessuno sapeva quanto ci fosse sotto.
Ora `-Xlint:deprecation` e `-Xlint:unchecked` sono attivi su tutti i
`JavaCompile`, e il codice che resta e' pulito con quei controlli accesi: se
domani rientra un uso deprecato, si vede subito invece di sedimentare.

Quasi tutto si e' chiuso con una soppressione mirata — 11 `@SuppressWarnings`
in Java, 8 `@Suppress` in Kotlin, 4 `@file:OptIn(DelicateCoroutinesApi)` — dove
l'API deprecata e' quella giusta e non ha un sostituto utilizzabile: le vibrazioni
prima di API 26, `AccessibilityNodeInfoCompat`, `KeyEvent.characters` per
`KEYCODE_UNKNOWN`, il giro `startActivityForResult`/`onActivityResult` di
`DecryptActivity`, che e' legacy da entrambi i lati e va tenuto in coppia.

Tre punti erano invece codice da correggere davvero:

- `AccessibilityUtils.shouldObscureInput` verificava che una costante `String`
  non fosse null — condizione sempre vera, e infatti il compilatore lo diceva;
- `CipherPanel` faceva `0x22FFFFFF.toInt()` su un letterale gia' `Int`, e
  `AdvancedScreen` `value.toInt().toString()` su un `Int`: due no-op;
- `CipherActions` chiamava `?.toString()` su una stringa.

Sul lato nativo la catena ndkBuild compila senza un warning su tutte e quattro
le ABI, con tre `-Wno-` aggiunti per famiglie che vengono dal codice di
HeliBoard e dall'NDK 28, non da noi: `deprecated-copy`,
`unused-but-set-variable`, `vla-cxx-extension`.

**Due cose lasciate apposta.** `LocaleUtils.constructLocale` continua a usare il
costruttore `Locale(...)` deprecato invece di `Locale.Builder`: su Android il
costruttore normalizza i codici ISO storici (`he` -> `iw`, `id` -> `in`) e il
Builder no, e l'app ha i sottotipi e il CSV dei dizionari scritti proprio in
`iw`. Cambiarlo avrebbe spostato il comportamento dentro una pulizia, che non e'
il posto giusto per farlo — quindi soppressione e commento che spiega il perche'.
E restano sette warning "nome non valido su Windows" su nomi di test di HeliBoard
che contengono `"` e `|`: quei caratteri sono esattamente cio' che il test
descrive, rinominarli renderebbe il nome falso.

Test unitari invariati rispetto a prima: 198, con gli stessi fallimenti
preesistenti (InputLogicTest e i due XLinkTest, che chiedono la rete).
`lintVitalRelease` verde.

+++ app/src/main/java/helium314/keyboard/cipher/CipherPanel.kt
@@ -99,7 +99,7 @@ object CipherPanel {
         // poggia semplicemente non si vede. Il bordo e' il colore del testo
         // molto trasparente, quindi si legge su chiaro e su scuro senza che
         // nessuno debba scegliere.
-        bolla?.setStroke(1.dp(trovato), colori.get(ColorType.KEY_TEXT) and 0x22FFFFFF.toInt())
+        bolla?.setStroke(1.dp(trovato), colori.get(ColorType.KEY_TEXT) and 0x22FFFFFF)
         trovato.findViewById<TextView>(R.id.cipher_panel_close)?.apply {
             setTextColor(colori.get(ColorType.KEY_TEXT))
             setOnClickListener { chiudi() }
